Minor: move batch spilling methods to `lib.rs` to make it reusable
